return 값 형변환

+++ java_project/FirstJava/src/MyProject/Calculator.java
@@ -3,10 +3,10 @@ package MyProject;
 import java.util.Scanner;
 
 public class Calculator {
-    public long plus(int num1, int num2){ return num1 + num2; }
+    public long plus(int num1, int num2){ return (long)num1 + num2; }
 
     public long minus(int num1, int num2){
-        return num1 - num2;
+        return (long)num1 - num2;
     }
 
     public long multi(int num1, int num2){
